FoodShop/requirements.txt
captcha==0.3
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-cors-headers==2.5.3
django-filter==2.1.0
django-redis==4.10.0
django-simpleui==2.0.4
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2
Pillow==6.0.0
pycryptodomex==3.7.2
PyJWT==1.7.1
python-alipay-sdk==1.10.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
